+++ sdk-core/src/test/java/io/nem/symbol/sdk/api/SecretLockSearchCriteriaTest.java
@@ -30,6 +30,7 @@ class SecretLockSearchCriteriaTest {
   void shouldCreate() {
     SecretLockSearchCriteria criteria = new SecretLockSearchCriteria(account1.getAddress());
     Assertions.assertNull(criteria.getOrder());
+    Assertions.assertNull(criteria.getSecret());
     Assertions.assertNull(criteria.getPageSize());
     Assertions.assertNull(criteria.getPageNumber());
     Assertions.assertNull(criteria.getOffset());
@@ -43,6 +44,7 @@ class SecretLockSearchCriteriaTest {
     criteria.setOrder(OrderBy.DESC);
     criteria.setPageSize(10);
     criteria.setPageNumber(5);
+    criteria.setSecret("someSecret");
     criteria.setOffset("abc");
     Assertions.assertEquals(account1.getAddress(), criteria.getAddress());
 
@@ -50,6 +52,7 @@ class SecretLockSearchCriteriaTest {
     Assertions.assertEquals(10, criteria.getPageSize());
     Assertions.assertEquals(5, criteria.getPageNumber());
     Assertions.assertEquals("abc", criteria.getOffset());
+    Assertions.assertEquals("someSecret", criteria.getSecret());
   }
 
   @Test
@@ -59,6 +62,7 @@ class SecretLockSearchCriteriaTest {
         new SecretLockSearchCriteria(account1.getAddress())
             .order(OrderBy.ASC)
             .pageSize(10)
+            .secret("someSecret")
             .pageNumber(5);
 
     criteria.offset("abc");
@@ -67,14 +71,17 @@ class SecretLockSearchCriteriaTest {
     Assertions.assertEquals(5, criteria.getPageNumber());
     Assertions.assertEquals("abc", criteria.getOffset());
     Assertions.assertEquals(account1.getAddress(), criteria.getAddress());
+    Assertions.assertEquals("someSecret", criteria.getSecret());
   }
 
   @Test
   void shouldBeEquals() {
 
-    SecretLockSearchCriteria criteria1 = new SecretLockSearchCriteria(account1.getAddress());
+    SecretLockSearchCriteria criteria1 =
+        new SecretLockSearchCriteria(account1.getAddress()).secret("someSecret");
 
-    SecretLockSearchCriteria criteria2 = new SecretLockSearchCriteria(account1.getAddress());
+    SecretLockSearchCriteria criteria2 =
+        new SecretLockSearchCriteria(account1.getAddress()).secret("someSecret");
 
     Assertions.assertEquals(
         new SecretLockSearchCriteria(account1.getAddress()),
@@ -90,7 +97,7 @@ class SecretLockSearchCriteriaTest {
     Assertions.assertNotEquals(criteria1, criteria2);
     Assertions.assertNotEquals(criteria1.hashCode(), criteria2.hashCode());
 
-    criteria2.pageNumber(100);
+    criteria2.secret("someSecret");
     Assertions.assertNotEquals(criteria1, criteria2);
     Assertions.assertNotEquals(criteria1.hashCode(), criteria2.hashCode());