+++ sdk-vertx-client/src/test/java/io/nem/symbol/sdk/infrastructure/vertx/RepositoryFactoryVertxImplTest.java
@@ -53,6 +53,7 @@ public class RepositoryFactoryVertxImplTest {
     Assertions.assertNotNull(factory.createHashLockRepository());
     Assertions.assertNotNull(factory.createSecretLockRepository());
     Assertions.assertNotNull(factory.createMultisigRepository());
+    Assertions.assertNotNull(factory.createFinalizationRepository());
     Assertions.assertNotNull(factory.createJsonSerialization());
 
     factory.close();